requirements.txt
boto3==1.26.114
psycopg2-binary==2.9.6
awscli-local==0.14
